\name{dowjones}
\alias{dowjones}
\alias{dowjones}

\title{Negative log-returns of DOW JONES.}

\description{Series of negative log-returns of the U.S. stock market index Dow Jones.}

\details{From the series of n = 8785 closing prices \eqn{S_{t}}, \eqn{t=1,2,...}, for the Dow Jones stock market index, recorded from January 29, 1985 to December 12, 2019, the series of negative log-returns.

\eqn{
X_{t+1} = - \log(S_{t+1}/S_t), \quad 1\leq t\leq n-1}{X_{t+1} = - log(S_{t+1}/S_t),   1<= t <= n-1}


is available. Hence the dataset (negative log-returns) contains \eqn{8784} observations.
}

\format{A \eqn{8784*2}{8784*2} data frame.}

\keyword{datasets}
